+++ examples/timestamps/models/models.gen.go
@@ -26,8 +26,10 @@ type PGClient struct {
 	// saw in the table we used to generate code. This means that you don't have to worry
 	// about migrations merging in a slightly different order than their timestamps have
 	// breaking 'SELECT *'.
-	rwlockForUser    sync.RWMutex
-	colIdxTabForUser []int
+	rwlockForUser                sync.RWMutex
+	colIdxTabForUser             []int
+	rwlockForGetUserAnywayRow    sync.RWMutex
+	colIdxTabForGetUserAnywayRow []int
 }
 
 // bogus usage so we can compile with no tables configured
